# docs/swagger.yaml
definitions:
  domain.Segment:
    properties:
      id:
        type: integer
      name:
        type: string
      percent:
        type: number
      ttl:
        type: string
    type: object
  domain.UpdateRequest:
    properties:
      segments_add:
        items:
          $ref: '#/definitions/domain.Segment'
        type: array
      segments_del:
        items:
          $ref: '#/definitions/domain.Segment'
        type: array
      user_id:
        type: integer
    type: object
info:
  contact: {}
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1.0"
paths:
  /api/segment:
    post:
      consumes:
      - application/json
      description: Create a new segment.
      parameters:
      - description: admin
        in: header
        name: User-role
        type: string
      - description: Segment attributes
        in: body
        name: segment_attrs
        schema:
          $ref: '#/definitions/domain.Segment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "403":
          description: Forbidden
      summary: create a new segment
      tags:
      - Segment
  /api/segment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a segment.
      parameters:
      - description: Segment ID
        in: path
        name: id
        required: true
        type: integer
      - description: admin
        in: header
        name: User-role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "404":
          description: Not found
      summary: delete a new segment
      tags:
      - Segment
  /api/user/get/{user_id}:
    get:
      description: Get list of user`s segments.
      parameters:
      - description: Segment ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No content
        "400":
          description: Bad request
        "404":
          description: Not found
      summary: Get list
      tags:
      - User
  /api/user/get_history:
    get:
      consumes:
      - application/json
      description: Get file with history of user`s segments
      parameters:
      - description: Segment ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Time start
        in: query
        name: start
        required: true
        type: integer
      - description: Time end
        in: query
        name: end
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: get file with history
      tags:
      - User
  /api/user/update:
    patch:
      consumes:
      - application/json
      description: Update user`s segments.
      parameters:
      - description: Update attributes
        in: body
        name: segment_attrs
        schema:
          $ref: '#/definitions/domain.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      summary: update user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"